Add the static assets and the most basic gin to use the template
Next steps are some back endy bits like persistence... not sure how
exactly I'm going to approach that persistence but it'll probably be
meta data in some sort of nosqly thing and the images on the file
system. Possibility to fake a back end by providing a database already
full of the meta data and some images to go with it are available.