Rip out boltdb and add sensible template rendering
Boltdb proved to be pretty useless for this case, we're pretty record
heavy in this type of application and I found I was just implementing
all the bits of a relational database in memory on a document. I'll
investigate alternative document databases once more to see if there's
some utility in them for this application. Failing that I'll put a
relational back end in, probably postgres.

Using a combination of:
 - https://elithrar.github.io/article/approximating-html-template-inheritance/
   (How to nest templates)
 - https://github.com/gin-contrib/multitemplate (Gin Package with the
   type of inheritance renderer specified above)
 - https://github.com/astaxie/NPWG_zh (Book: Network Programming with
   Go)

I've developed a better understanding of how html/template works (in
particular NPWG is worth a read if you want more information) and
created a sensible base template that can be used to add new pages
without redefining all the cruft like nav bars. Includes contains the
base templates you want to re-use and layouts apply their content to
these, I mention this because the naming convention was somewhat
un-intuitive to me but I've retained it because it was defined in the
approximating inheritance article. Not yet sure what's going to happen
when you have base.tmpl and anotherBase.tmpl but that's research to be
done when the need presents itself.