Fix archive page, add 404, bump bulma(css) fix relative links
Archive page and og headers now use the correct /post/$NUM URL for the
canonical post URL.

Add 404 page for missing post (will update gin to use this as the 404
all the time)

Bumped the bulma css library to latest version, this required some
fixes to the tile class names.

Fixed relative links from ./ to / this mistake worked fine on the
index page but not so great on /archive or /random!