Fix some routing errors, Add configuration with gcfg, Add pgsql based datasource
Fixed evaluation points of some functionality (inside the function request instead of inside the function creation) which was causing single time execution for the random endpoint.

Add configuration with gcfg, this allows the application to be flexibly configured by the odd.cfg file, see odd.cfg.sample for an example configuration which at the moment only covers the necessary properties to connect to a local pgsql database (which is what I ended up selecting, maybe we'll support multiple databases in the future).

Add the pgsql implementation for storing and retreiving posts, authentication still dummied out and at present no authorization at all.