Add most of authentication and user management
Got a little carried away here... all seemed related but this should probably have been a few smaller commits.

If there are no users then a Default user is created. If the Default user exists and is not deleted then its password will be changed on each invocation of owls-do-drive, this is to encourage application maintainers not to use the Default user but also to allow them to consider it a reasonably secure way of regaining control of a locked system without doing a database edit. (May add a switch in future that will restore the Default user if it is deleted to make this even easier.)

Passwords are salted and hashed with SHA1, salt is changed each time the password is changed.

Users authenticate themselves with the /api/token endpoint, automated for them by the /admin/login page, this provides them with a signed JWT token with their username as the only claim (since there are presently no permissions). The token is then set in a cookie because I couldn't work out how to ensure that non-api calls have it set as the Authorization header seems like the way this is solved in modern web is to do a one page app that "changes page" by an API call. (PRs welcome if you're reading this in the future).

Password resets work by mailing (logging at the moment, email to come) the recorded email address a reset token which is stored hashed with a salt on the server side when it is requested on /admin/forgot. Holders of the token can then use them as a one-time reset at /admin/reset which also logs them in.

/admin/users is a list of all the users which allows you to do the RUD of CRUD, C to come.
/admin/user/id allows you to edit the details of a user.

All of this to say that if you're willing to use the default user or create your own with the database you now have a working system. (Easiest thing to do is leave password and salt blank then use reset and check the log.)