Fix password reset
When the user attempts to reset their password the natural ordering
means that they can only use their oldest valid password reset.
Instead use the most recent password reset which is still valid. Not
perfect but certainly better and means the user can always request a
new password reset if they're unable to place their hands on a valid
reset token.

A better solution would be iterate over the valid password_reset
tokens and invalidate all but the N youngest tokens to ensure we're
not forced to do a lot of hashing.