Fix very tall images
Very tall images would simply drop off the end of the page because they were able to use so much width of the page to maintain aspect ratio.

Added an is-1 spacer on either side and limited the image to is-8 width in the grid. Spacers have the class "spacer" so they can be disabled later if we want a "compact" look. They are hidden on mobile but not because they are spacers, just because these particular spacers shouldn't show up there.