Fix users marked as deleted when edited
Users were being marked as deleted because the checkbox contained the checked
property. It was being set to "false" when the user was deleted but this does
not actually uncheck a checkbox it just sets the "checked" property to true.
The correct way to set based on the deleted status of the user is to
include/exclude the property when generating the page.

I'm not sure that I like the display:none checkboxes anymore. I think that the
page should just have the variables already in javascript or not even coming
out to the front end. After all, they're intended to be immutable!