Extract template/static asset directory paths to cfg
To support the use of 3rd party templates (and allow me to extract my proprietary content from the free repo) I've moved the hardcoded paths to variables in odd.cfg.

Additionally due to some weirdness with path construction previously the template directory had to be specified at least relatively (i.e. it wasn't implicitly in .) and it had to contain a trailing slash. This is no longer necessary if your templates directory is ./templates you can just set the property to templates.

If the data/assets directories do not exist then they will be created for you. This is not the case with the templates directory because some templates are required. A later version could sub in the generic template when a custom template is missing which would be a smoother user experience as new required pages are added or if their template only wants to override the front facing pages.