Prevent deletion of last user in database
This should prevent the last application user from becoming locked out
of the system. Of course this doesn't mean that the last application
user is still in possesion of thier email account so it is still
possible to get into a state where only a database edit can recover
admin access to the application. I think a switch which re-enables the
default admin would be wise.

Also the API probably shouldn't let you delete yourself.