Rationalise transactions
Remove the transaction where it was doing nothing and form a general
pattern for doing multistep transactions which might fail revolving
around the new transact function.

Fixes an issue in previous "Don't lock out the last user" function which
relied on high transaction isolation. The new approach not only doesn't
need a transaction but definitely won't broadcast an update unless the
update is acceptable.